requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pycparser==2.22
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3